+++ dali/pipeline/data/tensor_list.h
@@ -227,7 +227,7 @@ class DLL_PUBLIC TensorList : public Buffer<Backend> {
   /**
    * @brief Returns the number of tensors in the list.
    */
-  DLL_PUBLIC inline int ntensor() const {
+  DLL_PUBLIC inline size_t ntensor() const {
     return shape_.size();
   }
